Implement Charmander's STBL Recommendations (#5)
Implement's Charmander's recommendations, among other small fixes. Implements the ability for staff (mods and above) to bypass the STBL checks, as requested in the feature request for the STBL. 

Further, adds tests for searchtag.associate(), if only because this PR touches that function, and I want to be sure that it worked (and it caught a glitch in my code, so I feel justified adding it).

Squash-merge of 31 commits.