Minimize docker size II: alpine everything.
The images are a lot smaller, and more things are derived from the same
bage image, so that helps a lot. Moving the npm building into
weasyl-build and not doing it in weasyl-assets helps too because npm
building requires g++ for libsass, but not to actually run sass.

Also finally just chown root-owned things to weasyl too because
otherwise you have problems with volumes being owned by root.