Tidy up the last little node_modules issues.
The previous revert undid some things that needed to be kept, and this
chown _before_ calling VOLUME should solve be the end of it. Also, keep
node_modules and wheelhouse existing via .gitignore, so they don't have
to be recreated in ./wzl. Nothing in those directories will be
versioned, naturally.