Split out password auth step, and permit a space in TOTP codes.
From feedback in testing, seperately authenticate a user's password. Such compartmentalizes errors, and eliminates retyping a password and/or TOTP code if misentered.

From feedback in testing, altered the initialization TOTP code field lengths to length 7, permitting a space to be added, as some authenticators display codes seperated in groups of three (e.g., '123 456'). Strip the space out on the back-end via Python's 're' module and substituting spaces in the response for empty strings. Additionally applies this behavior to the two_factor_auth.verify() function, pre-stripping spaces from the tfa_response to permit the remainder of the code to work without modifications.