Opt for string.replace() vs. re.sub() (#15)
* Opt for string.replace vs. re.sub

The totp.verify() call in tfa.verify() will normally fail and return False if a space is in a TOTP code, as some 2FA apps like to put a space in the code. For instance, "123 456". The code would need to be provided as "123456" or the function would return False. Logically, the code is correct, technically, it is incorrect with the space. So tfa.verify() was altered to force the logically correct case to also be technically correct by force-stripping spaces. The space character is not a valid character in the recovery code charset. Add a test to make sure it functions as expected with a successful verification.