Shout notifications should link to /shouts/username#cid
For #212. Linking to /shouts#cid only works if the link is visited while logged in, as the target user to whom the shout was sent. Adding the user to whom the shout was sent permits a mote permanent link, and works if the link is copied and visited while logged or or sent to a different user if copied.