Incorporate Charmander's suggestions
Removes inclusion of email address in the templates and/or messages to the user, at least temporarially. Misaka needs to be brought current to display email addresses with double underscores.

Deduplicate code in some tests, opting instead to use the Bag() definition in utils.py

Alter flake8 config (setup.cfg) to ignore the 'templates/*' directory when running 'make check'; flake8 doesn't do native detection of HTML vs. Python (this has been present on prior commits, just not fixed until now).

Retool and cleanup forgotten password mechanisms to not prompt for username, and cleaned up controller to only send what is actually expected to be received (previously, the controller also attempted to get the year/month/day (but was never used by the forgotten password function); likely, was intended to be a check against the defined birthday stored in the account)

Provide email feedback to a user attempting to register a second account with an already registered email (provide information on how to change username if needed, or how to change password).