Add the ability for admins to view accounts pending creation
Provides a new admin-restricted page to view, and optionally purge, account creation requests in the logincreate table. In the event of an invalid logincreate record, the email wouldn't normally be displayed. So also add an invalid_email_addr field to the logincreate table to permit storing the originally provided email (since the email field is unique to the table).

Addresses the first and second proposals in #302.