Cleanup of code & fix cron.py
Sort records by username and cleanup leftover debug line. Additionally add stagetitle to the account confirmation link page and the pending account page. Add simple in-line style to provide for a 1em spacing between table rows.

Fix a long-standing issue with cron.py attempting to perform arithmetic operations against an Arrow object. Corrects code to use arrow.utcnow().timestamp instead of arrow.utcnow(), which would result in a TypeError if one attempted to perform an operation such as `arrow.utcnow() - (86400 * 2)`. As a consequence, this likely resulted in breaking of the SQLAlchemy transaction context manager, and resulted in a rollback of any operations performed inside of the context manager.