Tweak display of invalid records; remove stray anchor tag
Check the invalid Boolean first, then check if the invalid_email_addr has a value recorded. Once this has been live for two days, it should always have one set, but since this is a migration, displaying a record as-is will display the token in the email address field. Granted, it doesn't affect the ability to purge the record, but it isn't quite descriptive. So handle the set and unset cases for that field.

Assign a page title to the Pending Accounts page.