Only attempt to HTTP/2 server push critical resources once per session
Granted, according to the Cloudflare page on the topic of HTTP/2 server push, if the UA has the pushed content cached it should ignore the push and cancel the download. But if we know we've pushed the resources to the UA, let's avoid repushing (this assumes, of course, the pushed resources remain cached in the client, and the ephemeral cookie is properly cleared when the browser application is closed).