Replace lxml with html5lib
lxml is doing strange things to production, and etree is a terrible API. html5lib includes a sanitizer that’s mostly nice, except for the URL unescaping designed for a blacklist approach (I’m pretty sure it doesn’t introduce any vulnerabilities) and questionable approach to CSS filtering (validating colors only in `margin`/`padding`/…, which we bypass).

html5lib is said to be slower, but I wrote a useless and very inefficient escape checker for simplified user links that’s removed with this commit; it might balance out. If not, saving prerendered HTML in the database was probably a good idea anyway.

We should compare the results of each version on every piece of Markdown in the database before deploying this.