Don't delete records of rejected/flagged spam.
Currently, there's no mechanism built in to correctly clean up when a
record is deleted.

Alters the mechanism of the spam queue, such that hiding an item where
it is flagged as spam now omits it from the spam queue (in effect, this
serves as a 'reviewed' flag, if the item is marked as spam).

Merge the /deletespam/journal and /deletespam/submission routes down to
a single /spam/remove route (DRY, and all).

Adds a hide and unhidejournal function to ``moderation.py``, and updates
other hide/unhide functions to use the d.engine.execute() pattern over
the older d.execute().