Remove broken libweasyl.legacy.plaintext and its unused dependent; merge libweasyl.legacy.login_name and weasyl.define.get_sysname
libweasyl.legacy.plaintext didn’t do what it claimed to, and the Unicode-stripped password check in libweasyl.models.users was broken because of it – but it isn’t used in the current version of Weasyl anyway.

libweasyl.legacy.login_name was similar to weasyl.define.get_sysname, and intended to be used in the same situations. It’s better to have only one. It’s now reexported by weasyl.define (though unfortunately not explicitly, like via `__all__`), because it seemed like changing its references would be unnecessary noise.