Use a case-insensitive sort for ignored usernames
Slightly less surprising, although it still sorts punctuation according to ASCII.

We’re not on a version of PostgreSQL new enough for nondeterministic collations, but that’s okay, because usernames are supposed to be ASCII. (In practice, they’re sometimes not. Unicode wasn’t filtered out for some old accounts, and new not-normally-allowed usernames can be and have been created by admins. Also, ASCII here includes ASCII controls, for now – two existing usernames end with a tab.)