Remove collections graphic from upload pages
It’s kind of distracting, hard to read, advertises a feature that isn’t quite at the level it could be at, and takes up a lot of space – especially with narrow viewports.

The button on submission pages is a nice enough way to introduce it for now, I think.