Make profile folder previews show most recent, obey blacklists
Instead of showing a random submission from a random selection of three non-empty folders, show the most recently updated folders with the corresponding most recent updates.

Fixes #725, and behaviour when all submissions in a folder are friends-only.

The query is slower, but hopefully not slow enough to cause problems in practice.