Improve password reset
- always send an e-mail so people aren’t left guessing as to whether something is eventually coming (and so various behaviours are independent of whether an account with the given e-mail address exists on the site)

- look up e-mail case-insensitively (but prefer case-sensitive for backwards compatibility – there are 700+ e-mail addresses with multiple associated accounts because registration is case-sensitive! registration being `<input type="email">` but password reset being `<input type="text">` might have made this even more visible than usual on mobile.)

- look up e-mail at reset time so that a password reset request sent to an account’s previous address doesn’t work (even though the window was only one hour)

- drop 5-minute timer that starts on visiting the link, because it doesn’t do much for security and could be incredibly frustrating

- allow requesting a reset and performing the reset on different IP addresses, because people’s IP addresses can change and they can use multiple devices

- don’t require people to reenter their username and e-mail address, to make use of the password reset less annoying

- further to the previous, *show* the username so the feature can also act as a forgotten username lookup

- sidestep any potential side channel attacks on the token (and questions of whether keeping the tokens in the database is a risk) by hashing tokens and therefore not knowing them

- reduce the length and complexity of tokens, just in case someone needs to retype one

- log password resets

An important future improvement is to make this at least as resistant to abuse as the signup page, so we should start by applying the CAPTCHA to it as soon as the server is able to reach Google to validate them again.

One new risk this introduces is that someone with access to the mailbox will be able to view the owner’s Weasyl username undetected by them, but I think it’s acceptable.

Letting a user skip the login screen after changing their password would probably be helpful, but with 2FA and status checks having to be moved around to accomplish that, it seems best to leave that for later.