Set coverfile default value to empty string
The string 'boop' is not a decodable image format, and Sanpera/ImageMagick will raise a GenericMagickError if the coverfile form parameter is missing from the POST-ed page. Normally, this doesn't occur, as a selected, or unselected coverimage with either result in defining an image, or not, either way overwriting the coverfile parameter's value, and bypassing the check. However, it could be reached if the form parameter is deleted prior to submission (such as via the developer tools, or an intercepting proxy).