Remove start of optimized in-table media
This was enabled for a while in production and worked, but I never finished it and it isn’t covered by automated tests. It looks like it won’t work unchanged with Python 3, so now is a good time to remove it. We can try again after the language port and incremental improvements to the existing media tables (unique constraints on link types, removal of media-media links).