Fix @supports_json’s JSON path for Python 3
`json.dumps` now returns text, which `Response` doesn’t accept without a `charset`. It sets a `charset` of `None` when the `content_type` isn’t `text/*`, so the other uses of `Response(content_type=...)` in Weasyl are fine; this one broke because it was `application/*`.

`Response(json=...)` encodes as JSON, encodes as UTF-8, *and* sets a `Content-Type`.