Remove inter-media links
They add a lot of complexity along with other types of overhead, and are generated lazily in confusing ways while providing only a small benefit (not regenerating covers when identical submission files are reuploaded).

This slightly breaks the public HTTP API’s backwards-compatibility, but anything making use of the removed properties is already broken.