Remove `InputWrap`
I don’t think it’s necessary anymore…? WebOb’s code certainly gives the impression of trying to handle the situation itself, and I haven’t seen the corresponding event show up in Sentry. But if it is necessary, I’m pretty sure we shouldn’t do it like this.