Rely on `Origin` header for CSRF prevention
It’s supported by all modern browsers. Allows guests to be cookie-free and to share cache.

Some of the 2FA edge cases are probably broken (and it looks like they already were before), but I’d like to refactor that soon anyway.

SQLAlchemy is horrible as usual.

C++ compiler added (including to CI) to support SQLAlchemy 1.4.