Add new social site pairs to form when last one is filled
instead of using a second submit button.

I think this strikes a good balance of:

- not overriding the main Enter action of saving the form
- not causing uncertainty as to whether Enter will add a new field
- being convenient

Maybe ideally it would remove fields if the second-to-last pair became empty, though?

Also, it still doesn’t interact well with browser back/forward, sometimes getting jumbled when saved (because of the sorting applied by the server) and losing new entries past the first when not (because they’re created by a script). Not sure what the best way to solve that is.