Fix `strip_html` to return something closer to plain text
Even when I originally wrote this in [2013-09-25 :/Replaced BBCode with Markdown], the output of `strip_html` was used escaped, so it wasn’t correct then. Making the parser class local to the call with a `text_parts` class variable was also questionable style.

As of Python 3.5, `HTMLParser` never calls `handle_entityref` or `handle_charref` by default.