Simplify `options` parameter of common templates, and parameters of all templates
by having `options` be a non-`None` tuple at all times, and `argv` be non-`None` at all times.

`options` would be better as an instance of a class with separate fields for each option (and ideally not exist at all, making use of template inheritance), and `argv` is the one that’s *really* more appropriate as a tuple since it’s heterogeneous, but those are larger refactorings.