Add an explicit AI policy
I tried to match the surrounding tone/formatting, even though I think it isn’t ideal. All the policies could probably use editing and tables of contents, but… later! Also resisted the urge to specify, wordily, that everything here is up to reasonable intent-based interpretation according to site staff – but like the rest of the rules, it is. (Yes, art is also “a process based on the work of people other than the author”.)

Hopefully this makes it clearer to everyone that, in short, “AI art” isn’t allowed on Weasyl.