Always construct absolute URL for `Location` from return path parameter
Of course it was exploitable!

This is “fixed” in WebOb 1.8.8, but it’s not really WebOb’s responsibility to turn `Location: //…` into an absolute URL where the value is treated as a path-absolute URL. In fact, since RFC 7231 (the latest HTTP/1.1 specification), `Location` doesn’t need to be absolute; WebOb shouldn’t touch it at all. Treating the value differently from clients is bound to cause confusion (e.g. when we move away from WebOb, would this have been a problem waiting to happen?). WebOb’s current fix is to alter the prefix to `/%2f`, which is *semantically different*.

WebOb 1.8.8 also checks for the specific prefix `"//"`, but it still uses `urljoin`, which will happily cause exactly the same problem with `" //evil.example"` (substitute tab for space on older patches of Python, e.g. 3.10 before 3.10.12 (CVE-2023-24329)). That’s not an issue for what’s probably the most common way to have this open redirect vulnerability – what we did, echoing the request path – but, again, it’s not WebOb’s issue in general.

Finally, we should probably just redirect to slash-normalized URLs globally before even getting to the Python layer, but… another time.