Remove admin tool to get signup e-mail verification link
for now! It’s a bit hazardous:

- username field makes it easy to forget to check that the e-mail address is correct

- e-mail field accepts unexpected wildcards (`%` too, but especially `_`, could be abused in a support request)

- link is presented to admin as clickable, and just loading it performs the action (particularly bad on platforms where the gestures to copy a link and preview it are similar or the same – verification on `GET` is a bug)

- retrieved link isn’t distinct from normal link, making misuse harder to trace

In general, safe and correct use for support isn’t documented anywhere, which is a common problem with these tools. (But I’m still handling support for the moment.)

This tool is also rarely used, and I don’t want to deal with it during the upcoming changes to signup; it might come back in an improved form after that.