Use common criteria for determining whether a key can act as a shortcut
(and common implementation.)

This allows F and C shortcuts to work even when a link is focused, for example.

Adding `<select>` is particularly important now that letters are in, since that’s a way of interacting with that control. Are there any other interactive elements that our shortcuts conflict with…?

Also, the page isn’t supposed to scroll horizontally, but `preventDefault` for that and general correctness when using the arrow keys.