Show all search form controls initially; group type-specific fields into fieldsets
`search.js` controls visibility of type-specific fields depending on what type is selected, but when it fails to do that, it doesn’t make sense for the fallback to depend on what search led to the form: the form allows the user to pick a new type. Instead, indicate which sets of fields apply to which types.