Replace `last-input` bottom padding with top padding in search form
`last-input` is hard to maintain across all the combinations of controls in different form states.

`form-actions` creates less space than `last-input` and looks off for now, but this is because the thick line beneath `.input` elements is a box-shadow that doesn’t reserve space for itself, which I plan to fix soon.