dev: Implement watch mode basics
I started out with a solution based on merging a dynamic set of async iterables to implement working recursive watch in Node, but eventually, switching to Deno just seemed nicer. There are some leftovers from that direction in this implementation. All of the libraries (Node’s `fs`, chokidar on npm, Deno’s `Deno.watchFs`) for watching seem… bad, which is understandable for a complicated cross-platform problem (the filesystem itself is a bad interface!), but I really just wanted guaranteed inotify with a minimum of layers of confusion.