Use `ADD --checksum` to fetch source releases of mozjpeg and ImageMagick
It looks like using `--checksum` stops the builder from rechecking the URL on every build, which was the previous problem with `ADD`. `--link` makes the layer cacheable even when previous instructions change (I would hope Docker would cache the file itself with `--checksum` alone), and in theory this is more parallelizable too.