Work around pip 23.0.1 dependency resolution bug
pip is resolving a Poetry dependency to a version that isn’t the one we have pinned. Luckily, `--require-hashes` still catches it.

Both `-c requirements.txt` and `--no-deps` can work around the issue.

This was fixed at some point between pip 23.0.1 and the current pip 25.1.1, but I don’t want to add an additional Poetry-style lockfile for a pip upgrade and further bloat the image besides.

We should be able to upgrade to a fresher base image soon enough anyway.