Simplify thumbnail layout algorithm to remove thumbnail ratio and increase minimum layout width
As mentioned in the removed comment, this results in layouts that are only off what the original algorithm considers optimal by a small margin. On the subjective side, I think the results are actually better, with fewer single-thumbnail rows (but there are further improvements to be made when it comes to making layouts more balanced).