Remove director e-mail blacklist; match against blocked domains as a list of suffixes directly
The non-director-facing functionality of the director e-mail blacklist is the same as `disposable-domains.txt`. In practice, maintaining different block reasons under different source files and merging them into one configuration file at deploy time is more convenient than using the web UI. The list doesn’t change so often that syncing it and `SIGHUP`ing Gunicorn is an inconvenience.

This also switches to performing suffix lookups against the list directly, instead of relying on a public suffix list which…

- needs to be maintained according to the usual dependency considerations despite most updates being to its data (although in practice, we don’t need to keep it up to date)

- introduces complexity and data we never use (it was possible to attempt to block a domain and have that quietly achieve nothing because of the public suffix list)

- looks like it might have a questionable O(n²)-time implementation? (but we should be limiting the length of user input here anyway)

I stopped myself from implementing this as a binary search in an `mmap`ed file of reversed names to save ~23 MB of memory, but you have to admit it’s pretty unsatisfying that every worker process loads a separate instance of this static data on startup.