Replace generic profile media route with avatar-only redirect
As far as I know, this functionality is only used internally by Markdown, for avatars (and I have plans for rich text that will remove this use). `X-Accel-Redirect` has the downside of being unable to share cache with the canonical media path, and we don’t actually want to expose `avatar-source`.

Notable behaviour changes:

- avatar links to nonexistent users now display a broken image instead of the default avatar
- the response now has a nonzero cache freshness time
- the response is now a redirect, so a nice URL won’t be preserved if a user navigates directly to the route (should be fixed in the future by making the canonical media URL itself nice)
- `banner` is no longer accessible

The result has a lot in common with `/api/useravatar`.

Partially reverts “Generalize media link URLs” (2014-01-10).