Remove generic submission media route
- It exposes media links that shouldn’t be exposed (`thumbnail-source`).
- It’s inefficient, using an `X-Accel-Redirect` to serve the file through nginx instead of serving an HTTP redirect to a resource that might already exist in cache, creating redundant long-lived cache entries (the response `Cache-Control` is the in-nginx redirect target’s, not the `max-age=0` that the upstream serves).
- It doesn’t work for friends-only submissions!
- Its broad pattern matches all sorts of URL accidents and serves up success responses. (Many of them are unintended relative links from within submission descriptions, which #1610 should help with from another angle over time.)
- I can’t find any legitimate attempted uses of it in our request logs.

This ports the more specific nginx rule from production to the dev Docker Compose setup.

A previously-known related problem that remains after the removal: the more restrictive route still allows any file to be served with any filename (regardless of whether it’s linked to the submission in the URL, and regardless of the actual link types that reference it), with the same cache issues. The long-planned upcoming media and post route changes will fix this by tying all media to submissions (`/~user/posts/id-slug-thumb.avif`) and redirecting requests with the wrong filename, so I’m not going any farther for now.